force reload after approe/deny
